move collection_shard_distribution into separate file
