+++ lib/collection/tests/common/mod.rs
@@ -1,7 +1,8 @@
 use collection::config::{CollectionConfig, CollectionParams, WalConfig};
 use collection::operations::types::CollectionError;
 use collection::optimizers_builder::OptimizersConfig;
-use collection::{ChannelService, Collection, CollectionId, CollectionShardDistribution};
+use collection::shard::collection_shard_distribution::CollectionShardDistribution;
+use collection::{ChannelService, Collection, CollectionId};
 use segment::types::Distance;
 use std::num::NonZeroU32;
 use std::path::Path;
