+++ lib/collection/tests/snapshot_test.rs
@@ -1,7 +1,8 @@
 use crate::common::TEST_OPTIMIZERS_CONFIG;
 use collection::config::{CollectionConfig, CollectionParams, WalConfig};
+use collection::shard::collection_shard_distribution::CollectionShardDistribution;
 use collection::shard::Shard;
-use collection::{ChannelService, Collection, CollectionShardDistribution};
+use collection::{ChannelService, Collection};
 use segment::types::Distance;
 use std::num::NonZeroU32;
